Move measures geo areas forms into geo areas folder
